+++ R/in-notin.R
@@ -8,6 +8,11 @@
 #' @param x vector or NULL
 #' @param y vector or NULL
 #'
+#' @details
+#' Falls back to base::`%in%` when x and y don't share a common type.
+#' This means that the behaviour of base::`%in%` is preserved (e.g. "1" %in% c(1, 2) is TRUE)
+#' but loses the speedup provided by vec_in.
+#'
 #' @export
 #'
 #' @examples
@@ -22,12 +27,19 @@
 '%in%' <- function(x, y) {
   if (is.character(x) && is.character(y)) {
     x %chin% y
-  } else if (is.list(y)) {
-    # https://github.com/markfairbanks/tidytable/issues/565
-    base::'%in%'(x, y)
-  } else {
+  } else if (vec_ptype_compatible(x, y)) {
     vec_in(x, y)
+  } else {
+    base::'%in%'(x, y)
   }
+}
+
+#' Check if two vectors have compatible ptypes
+#' @noRd
+vec_ptype_compatible <- function(x, y) {
+  # https://github.com/markfairbanks/tidytable/issues/565
+  # https://github.com/markfairbanks/tidytable/issues/632
+  tryCatch({vec_ptype_common(x, y); TRUE}, error = function(e) FALSE)
 }
 
 #' @export