+++ tests/testthat/test-in-notin.R
@@ -4,6 +4,10 @@ test_that("%in% works", {
   expect_equal(c("a", "d") %in% c("a", "b"), c(TRUE, FALSE))
 })
 
+test_that("%in% falls back to base implementation on arguments that don't share a prototype", {
+  expect_equal(c("1", "4") %in% c(1, 2), c(TRUE, FALSE))
+})
+
 test_that("properly handles character comparison with NA", {
   expect_equal(c("a", "b") %in% NA, c(FALSE, FALSE))
 })